Correction to my earlier update regarding warning
In my previous edit, I eliminated the calls to CharToLetter on lines 143-147, but I just debugged and figured out what's needed is to change the calls to LetterToChar.  Once I compiled that, I was able to figure out why I was getting the warnings in the first place.  There's another bug on line 114 that causes the last column to never get set.  This isn't that big of a deal when the matrix being read in has an ambiguous column/row for amino acid character 'X', but it makes a big difference when the custom matrix being supplied contains no rows/columns for ambiguous characters.  Thus, the sanity check code prints a single warning on line 142.

I also removed the ExitLoop code so that all the warnings are printed (which was helpful for debugging why the matrix was generating warnings).